refactor: ExpenseType 파일 별도 분리
